Update unzip_pdb_files_commented.py
The provided code is designed to assist in the extraction and analysis of data from the RSCB PDB (Protein Data Bank). The PDB is a widely used repository of experimentally determined 3D structures of biological macromolecules, including proteins and nucleic acids.

The code begins by importing the necessary modules: os for operating system-related functions, gzip for handling compressed gzip files, and shutil for file operations.

Next, the code sets the input and output directories. The data_dir variable should be modified to reflect the path to the directory where the RSCB PDB files are located. Similarly, the output_dir variable should be set to the desired location for the unzipped output files.

To ensure that the output directory exists, the code uses the os.makedirs() function with the exist_ok=True argument. This function creates the directory if it does not already exist or does nothing if it does.

The code then searches for PDB files in the input directory using the os.listdir() function. It filters the files based on the condition that they end with the extension ".ent.gz". The resulting list of PDB files is stored in the pdb_files variable.

Next, the code enters a loop to process each PDB file. For each file, it performs the following steps:

Opens the gzip file using gzip.open() with the full path of the file.
Opens a new file in write binary mode using the open() function with the corresponding file name in the output directory. The ".gz" extension is removed from the file name by using pdb_file[:-3].
Copies the contents from the input gzip file to the output file using the shutil.copyfileobj() function. This function efficiently copies data from one file-like object (f_in) to another (f_out).