Create clean_pdb_data.py
The code then proceeds to remove rows from the DataFrame that have empty values in a specific column. By using the dropna() function and specifying the subset as the column at index 11 (12th column in zero-based indexing), all rows with empty values in this column are eliminated. This step helps ensure data consistency and quality by removing incomplete or missing entries.

Next, the code removes duplicate rows from the DataFrame based on values in the second column (index 1, 2nd column in zero-based indexing). The drop_duplicates() function is employed to identify and remove duplicate rows, allowing for streamlined analysis and avoiding redundancy in the dataset.

Finally, the resulting DataFrame, which now contains the cleaned and preprocessed data, is saved back to the "pdb_data.csv" file, overwriting the original file. The data is written in the CSV format with the specified delimiter as ';' and the index=False argument ensures that the row index is not included in the saved file.